PoC ready for early PR review
