+++ src/WebProducer/Vinyl-zip.js
@@ -3,8 +3,6 @@
 // Third party dependencies (Typically found in public NPM packages)
 const through = require("through2");
 const Vinyl = require("vinyl");
-
-// Project dependencies
 const zip = new require("jszip")();
 
 /**
