+++ src/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "debug": "node --inspect-brk=9229 ./node_modules/.bin/serverless offline"
   },
-  "devDependencies": {
+  "dependencies": {
     "aws-sdk": "^2.594.0",
     "gulp-htmlmin": "^5.0.1",
     "gulp-minify-css": "^1.2.4",
