--- test/www.tforster.com.postman_collection.json
@@ -1,87 +1,0 @@
-{
-	"info": {
-		"_postman_id": "d7cbf847-f139-4951-afac-034dfda038d3",
-		"name": "www.tforster.com",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "Get pages",
-			"request": {
-				"auth": {
-					"type": "bearer",
-					"bearer": [
-						{
-							"key": "token",
-							"value": "{{jwt}}",
-							"type": "string"
-						}
-					]
-				},
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "graphql",
-					"graphql": {
-						"query": "query{\r\n  pages\r\n  {\r\n      \r\n      Template\r\n      id\r\nBody \r\n      products {\r\n          Title\r\n          Teaser\r\n          Description\r\n      }\r\n          \r\n      \r\n  }\r\n}",
-						"variables": ""
-					}
-				},
-				"url": {
-					"raw": "http://localhost:1337/graphql",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "1337",
-					"path": [
-						"graphql"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Login",
-			"event": [
-				{
-					"listen": "test",
-					"script": {
-						"id": "1473ec92-4c64-4464-afaf-f3af0d646dfc",
-						"exec": [
-							"const response = pm.response.json();",
-							"",
-							"if(response && response.data && response.data.login && response.data.login.jwt) {",
-							"  pm.environment.set(\"jwt\", response.data.login.jwt);    ",
-							"}"
-						],
-						"type": "text/javascript"
-					}
-				}
-			],
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "graphql",
-					"graphql": {
-						"query": "mutation {\r\n  login(\r\n    input: { identifier: \"{{username}}\", password: \"{{password}}\", provider: \"local\" }\r\n  ) {\r\n    jwt\r\n    user {\r\n      \r\n      username\r\n    }\r\n  }\r\n}",
-						"variables": ""
-					}
-				},
-				"url": {
-					"raw": "http://localhost:1337/graphql",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "1337",
-					"path": [
-						"graphql"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
